Remove deprecated methods, move PropertyResolver
